Use Promise wrapper for file system interactions
Errors while reading and writing the namespace were not passed on to the
Promise chain, so the script would report that it was complete prior to
actually finishing file system interactions (where a further error may
then occur after we gave the all-clear). This commit wraps the callback-
based filesystem interaction in a Promise and passes errors to reject().

Chalk is also added to the project dependencies so that we don't rely on
the react-scripts dependency list.